requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.2.0
Flask-Mail>=0.10.0
Flask-Limiter>=3.5.0
Flask-Caching>=2.1.0
Werkzeug>=3.0.0
itsdangerous>=2.1.0
python-dotenv>=1.0.0
email-validator>=2.0.0
WTForms>=3.0.0
SQLAlchemy>=2.0.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
google-genai>=1.0.0
trafilatura>=1.6.0
pdfplumber>=0.11.0
razorpay>=1.4.0
python-magic>=0.4.27
Pillow>=10.0.0
celery>=5.3.0
redis>=5.0.0
pytest>=8.0.0
pytest-flask>=1.3.0
flower>=2.0.0
markdown>=3.5.0
pygments>=2.17.0
